cjoin

+++ src/scheduler.c
@@ -12,6 +12,9 @@
 
 int did_init = 0;
 
+// There is no need for it to be public, so I wont add to header
+int scheduler_free_waiting_thread(int tid);
+
 int scheduler_initialize_queues() {
     if (DEBUG) printf("Start: %s\n", __FUNCTION__);
 
@@ -19,16 +22,22 @@ int scheduler_initialize_queues() {
 	ready_medium = malloc(sizeof(FILA2));
 	ready_high = malloc(sizeof(FILA2));
 	executing = malloc(sizeof(FILA2));
+	blocked = malloc(sizeof(FILA2));
+	joined = malloc(sizeof(FILA2));
 	
 	if (ready_low == NULL) return MALLOC_ERROR;
 	if (ready_medium == NULL) return MALLOC_ERROR;
 	if (ready_high == NULL) return MALLOC_ERROR;
 	if (executing == NULL) return MALLOC_ERROR;
+    if (blocked == NULL) return MALLOC_ERROR;
+    if (joined == NULL) return MALLOC_ERROR;
 
 	if (CreateFila2(ready_low) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
 	if (CreateFila2(ready_medium) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
 	if (CreateFila2(ready_high) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
 	if (CreateFila2(executing) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+    if (CreateFila2(blocked) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+    if (CreateFila2(joined) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
 
     if (DEBUG) printf("End: %s\n", __FUNCTION__);
 
@@ -77,7 +86,7 @@ int scheduler_init() {
 /**
  * Create a context
  */
-int scheduler_create_context(ucontext_t* context, ucontext_t* next) {
+int scheduler_create_context(ucontext_t *context, ucontext_t *next) {
     if (DEBUG) printf("Start: %s\n", __FUNCTION__);
 
 	//Laura not sure se isso tbm nao deveria estar no fim da funcao... mas esta funcionando..
@@ -100,6 +109,34 @@ int scheduler_create_context(ucontext_t* context, ucontext_t* next) {
     return SUCCESS_CODE;
 }
 
+int scheduler_add_to_blocked_queue(TCB_t *thread) {
+    if (blocked == NULL) return  NULL_POINTER;
+    if (AppendFila2(blocked, thread) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+
+    thread->state = PROCST_BLOQ;
+
+    return SUCCESS_CODE;
+}
+
+int scheduler_remove_from_blocked_queue(TCB_t *thread) {
+
+    if (blocked == NULL) return  NULL_POINTER;
+    if (FirstFila2(blocked) != SUCCESS_CODE) return EMPTY_LINE;
+
+    do {
+        TCB_t *iteration_thread = (TCB_t *) GetAtIteratorFila2(blocked);
+
+        if (iteration_thread->tid == thread->tid) {
+            if (DeleteAtIteratorFila2(blocked) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+            thread->state = PROCST_APTO;
+            return SUCCESS_CODE;
+        }
+
+    } while (NextFila2(blocked) == SUCCESS_CODE);
+
+    return INVALID_BLOCKED_THREAD;
+}
+
 int scheduler_block_thread(csem_t *sem) {
 
 	if (DEBUG) printf("Start: %s\n", __FUNCTION__);
@@ -110,12 +147,13 @@ int scheduler_block_thread(csem_t *sem) {
 	TCB_t *executing_thread = scheduler_get_executing_thread();
 	if (executing_thread == NULL) return NULL_POINTER;
 
-	executing_thread->state = PROCST_BLOQ;
+	int add_result = scheduler_add_to_blocked_queue(executing_thread);
+	if (add_result != SUCCESS_CODE) return add_result;
 	
 	if (DeleteAtIteratorFila2(executing) != SUCCESS_CODE) return LINE_OPERATION_ERROR;		
 	if (AppendFila2(sem->fila, (void *)executing_thread)) return LINE_OPERATION_ERROR;
-	
-	return scheduler_schedule_next_thread(&(executing_thread->context));
+
+    return scheduler_schedule_next_thread(&(executing_thread->context));
 }
 
 int scheduler_free_thread(csem_t *sem) {
@@ -125,7 +163,9 @@ int scheduler_free_thread(csem_t *sem) {
 	
 	if (FirstFila2(sem->fila) != SUCCESS_CODE) return EMPTY_LINE;
 	TCB_t *thread_to_wake = sem->fila->first->node;
-	thread_to_wake->state = PROCST_APTO;
+
+	int remove_result = scheduler_remove_from_blocked_queue(thread_to_wake);
+	if (remove_result != SUCCESS_CODE) return remove_result;
 	
 	if (DeleteAtIteratorFila2(sem->fila) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
 	return scheduler_insert_in_ready(thread_to_wake);
@@ -210,16 +250,14 @@ int scheduler_schedule_next_thread(ucontext_t *context_to_leave) {
         scheduler_show_state_queues();
 		printf("\n - The thread that will be executed: %d\n", next->tid);
 	}
-	
-	if (context_to_leave == NULL) {
-	
-		return setcontext(&(next->context));
-	
-	} else {
-	
-		return swapcontext( context_to_leave, &(next->context) );
-	}
-    
+
+    if (context_to_leave == NULL) {
+        return setcontext(&(next->context));
+
+    } else {
+
+        return swapcontext(context_to_leave, &(next->context));
+    }
 }
 
 int scheduler_kill_thread_from_exec() {
@@ -228,9 +266,11 @@ int scheduler_kill_thread_from_exec() {
 	if (executing_thread == NULL) return NULL_POINTER;
 	
 	executing_thread->state = PROCST_TERMINO;
-	
 
 	if (DeleteAtIteratorFila2(executing) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+
+	int free_result = scheduler_free_waiting_thread(executing_thread->tid);
+	if (free_result != SUCCESS_CODE) return free_result;
 
 	// TODO: Test this somehow
 	free(executing_thread); // Laura does not have sure
@@ -249,9 +289,8 @@ ucontext_t *scheduler_send_exec_to_ready() {
 
 	int insert_ready_result = scheduler_insert_in_ready(executing_thread);
 	if (insert_ready_result != SUCCESS_CODE) return NULL;
-	
+
 	return &(executing_thread->context);
-	
 }
 
 
@@ -274,12 +313,92 @@ int scheduler_insert_in_ready(TCB_t *thread) {
     }
 }
 
-TCB_t* scheduler_get_executing_thread() {
+TCB_t *scheduler_get_executing_thread() {
     if (executing == NULL) return NULL;
     if (FirstFila2(executing) != SUCCESS_CODE) return NULL;
 
     return (TCB_t *) GetAtIteratorFila2(executing);
 }
+
+// Start cjoin functions
+
+JP_t *scheduler_get_pair_with_blocker(int tid) {
+    // If the queue is empty, there is no one waiting
+    if (FirstFila2(joined) != SUCCESS_CODE) return NULL;
+
+    do {
+        JP_t *join_pair = (JP_t *) GetAtIteratorFila2(joined);
+        if (join_pair->blocker_tid == tid) return join_pair;
+
+    } while (NextFila2(joined) == SUCCESS_CODE);
+
+    return NULL;
+}
+
+int scheduler_queue_has_thread(PFILA2 queue, int tid) {
+    if (FirstFila2(queue) != SUCCESS_CODE) return INVALID_THREAD;
+
+    do {
+        TCB_t *thread = (TCB_t *) GetAtIteratorFila2(queue);
+        if (thread->tid == tid) return SUCCESS_CODE;
+
+    } while (NextFila2(queue) == SUCCESS_CODE);
+
+    return INVALID_THREAD;
+}
+
+int scheduler_thread_exists(int tid) {
+    if (scheduler_queue_has_thread(ready_high, tid) == SUCCESS_CODE ||
+        scheduler_queue_has_thread(ready_medium, tid) == SUCCESS_CODE ||
+        scheduler_queue_has_thread(ready_low, tid) == SUCCESS_CODE ||
+        scheduler_queue_has_thread(blocked, tid) == SUCCESS_CODE) {
+        return SUCCESS_CODE;
+
+    } else {
+        return INVALID_THREAD;
+    }
+}
+
+int scheduler_wait_thread(int tid) {
+    TCB_t *executing_thread = scheduler_get_executing_thread();
+    if (executing_thread == NULL) return NULL_POINTER;
+
+    if (executing_thread->tid == tid) return CANNOT_JOIN_ITSELF;
+
+    JP_t *new_pair = malloc(sizeof(JP_t));
+    if (new_pair == NULL) return MALLOC_ERROR;
+
+    new_pair->blocker_tid = tid;
+    new_pair->blocked_thread = executing_thread;
+
+    if (FirstFila2(executing) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+    if (DeleteAtIteratorFila2(executing) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+    if (joined == NULL) return NULL_POINTER;
+    if (AppendFila2(joined, new_pair)) return LINE_OPERATION_ERROR;
+
+    int add_result = scheduler_add_to_blocked_queue(executing_thread);
+    if (add_result != SUCCESS_CODE) return add_result;
+
+    return scheduler_schedule_next_thread(&(executing_thread->context));
+}
+
+int scheduler_free_waiting_thread(int tid) {
+    JP_t *waiting_pair = scheduler_get_pair_with_blocker(tid);
+    if (waiting_pair == NULL) return SUCCESS_CODE;
+
+    if (DeleteAtIteratorFila2(joined) != SUCCESS_CODE) return LINE_OPERATION_ERROR;
+
+    TCB_t *waiting_thread = waiting_pair->blocked_thread;
+
+    free(waiting_pair);
+
+    int remove_result = scheduler_remove_from_blocked_queue(waiting_thread);
+    if (remove_result != SUCCESS_CODE) return remove_result;
+
+    return scheduler_insert_in_ready(waiting_thread);
+}
+
+// End cjoin functions
 
 void scheduler_show_state_queues() {
